fix(atribuicao): os extratores voltam para a pasta do canal, e o post orgânico deixa de virar anúncio
Consertos de triagem em cima do PR #263 do @brunomaiasl. A feature, o desenho da
função atômica no banco, a guarda de primeiro toque e o merge de
`source_metadata` ficam exatamente como ele escreveu.

1. **`lint:channels` (passo do `verify`, required check) reprovava.**
   `lib/leads/atribuicao-de-anuncio.ts` é camada de FEATURE e nomeava provider
   seis vezes, inclusive no nome exportado. O invariante 1 da doutrina de
   restrição de canal diz que nenhuma feature nomeia um provider — e a
   allowlist do lint é `lib/channels/` e `lib/waha/`, que são o próprio canal.
   O arquivo virou três, sem mudar uma linha de lógica:
     - `lib/leads/atribuicao-de-anuncio.ts` — tipos + gravação (agnóstico)
     - `lib/channels/atribuicao-de-anuncio-oficial.ts` — extrator do `referral`
     - `lib/waha/atribuicao-de-anuncio.ts` — extrator do payload cru
   `pnpm lint:channels` → "ok (60 arquivos de dívida conhecida, nenhum novo)".

2. **O extrator do transporte por QR não filtrava `sourceType !== "ad"`, e o
   irmão da API oficial filtrava.** Post ORGÂNICO compartilhado não é anúncio
   pago: sem o filtro, o primeiro compartilhamento carimba `meta_ads` no
   contato — e a guarda de primeiro toque, que é a parte boa do desenho, torna
   isso irreversível pelo caminho normal. Caso novo no teste, sabotado para
   provar que vigia: tirando o filtro, 1 de 13 falha; com ele, 13/13.

3. **`console.error` → `logger.error`** com contexto (`contact_id`, plataforma,
   detalhe truncado). O DoD proíbe console em código mergeado, e a falha
   continua tolerada de propósito — perder a atribuição não pode derrubar a
   entrada da mensagem — mas agora ela aparece.

4. **Renumerado 0161 → 0164** (arquivo, timestamp, cabeçalho, MANIFEST e o
   rótulo do bloco no baseline), e a linha do MANIFEST voltou para a ordem
   cronológica. Cinco PRs abertos disputavam o 0161.

Uma coisa que ele acertou e quase ninguém acerta: o bloco do apêndice **já
estava antes** da varredura de `anon` do baseline, e os dois `revoke ... from
public, anon` do item 9 do CLAUDE.md estavam lá desde o primeiro commit.
